requirements.txt
Django==6.0.4
django-jazzmin==3.0.4
python-dotenv==1.2.2
psycopg==3.3.3
Pillow==10.4.0

